requirements.txt
aiohttp
aiosignal
async-timeout
attrs
certifi
charset-normalizer
exceptiongroup
fire
frozenlist
idna
iniconfig
multidict
openai
packaging
pluggy
pytest
pytest-mock
python-dotenv
requests
ruff
six
termcolor
tomli
tqdm
urllib3
yarl
ollama
